requirements.txt
Django==4.1.7
djangorestframework==3.14.0
face_recognition==1.3.0
geocoder==1.38.1
numpy==1.22.3
opencv_contrib_python==4.5.5.64
opencv_python==4.7.0.72
overpy==0.6
pandas==1.4.1
Pillow==9.0.1
Pillow==9.5.0
playsound==1.3.0
Requests==2.30.0
scipy==1.8.1
torch==1.8.1+cu111
ultralytics==8.0.105
~pencv_python==4.5.5.64
